requirements.txt
# FastAPI Core
fastapi
uvicorn[standard]

# Database (Async)
sqlalchemy[asyncio]
asyncpg # PostgreSQL driver
aiosqlite # For testing

# Pydantic Settings
pydantic-settings

# Security
python-jose[cryptography]
passlib[bcrypt]

# HTTP Client for LLM
httpx

# Testing (Mandatory)
pytest
pytest-asyncio

alembic